third_party/blink/web_tests/fast/css/getComputedStyle/getComputedStyle-border-image-slice-expected.txt

Test to make sure border-image-slice is correctly parsed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS computedStyle.getPropertyValue('border-image-slice') is '10'
PASS computedStyle.getPropertyValue('border-image-slice') is '30%'
PASS computedStyle.getPropertyValue('border-image-slice') is '10'
PASS computedStyle.getPropertyValue('border-image-slice') is '10 30%'
PASS computedStyle.getPropertyValue('border-image-slice') is '30%'
PASS computedStyle.getPropertyValue('border-image-slice') is '10'
PASS computedStyle.getPropertyValue('border-image-slice') is '30% 10 10'
PASS computedStyle.getPropertyValue('border-image-slice') is '10 30%'
PASS computedStyle.getPropertyValue('border-image-slice') is '30% 30% 30% 10'
PASS computedStyle.getPropertyValue('border-image-slice') is '10'
PASS computedStyle.getPropertyValue('border-image-slice') is '30% 30% 30% 10'
PASS computedStyle.getPropertyValue('border-image-slice') is '30%'
PASS computedStyle.getPropertyValue('border-image-slice') is '30% fill'
PASS computedStyle.getPropertyValue('border-image-slice') is '10 fill'
PASS computedStyle.getPropertyValue('border-image-slice') is '2 4 8% 16% fill'
PASS computedStyle.getPropertyValue('border-image-slice') is '30% fill'
PASS computedStyle.getPropertyValue('border-image-slice') is '10 fill'
PASS computedStyle.getPropertyValue('border-image-slice') is '2 4 8% 16% fill'
PASS computedStyle.getPropertyValue('border-image-slice') is '100%'
PASS computedStyle.getPropertyValue('border-image-slice') is '100%'
PASS computedStyle.getPropertyValue('border-image-slice') is '100%'
PASS successfullyParsed is true

TEST COMPLETE
